+++ src/Candoumbe.Pipelines/Components/IHaveGitVersion.cs
@@ -10,12 +10,12 @@ namespace Candoumbe.Pipelines.Components;
 /// Marks a pipeline that manages a Solution that follow GitVersion
 /// </summary>
 [NuGetPackageRequirement("gitversion.tool")]
-public interface IHaveGitVersion : INukeBuild
+public interface IHaveGitVersion : INukeBuild, IRequireNuGetPackage
 {
     /// <summary>
     /// The GitVersion tool that can be used to version the project
     /// </summary>
-    [GitVersion(Framework = "net5.0", NoFetch = true)]
+    [GitVersion(Framework = "net5.0")]
     [Required]
     public GitVersion GitVersion => TryGetValue(() => GitVersion);